refactor(inventory): extract encumbrance calc into its own module
Move selectEncumbrance / significantTreasure / encBarStops (+ internal
fillTier, basicArmorTier, TIER_STATUS) out of the bloated inventory.ts into
encumbrance.ts (comment-free); inventory.ts re-exports them so no consumer
import changes. 517 → 334 lines.
