+++ src/OscSheet/features/inventory/inventory.ts
@@ -1,12 +1,9 @@
-import type { CharacterEncumbrance } from "@ose-foundry-core/types";
-import type { OSEActor, OseItem } from "@domain/types";
+import type { OseItem } from "@domain/types";
 import type {
   InventorySortKey,
   SortDir,
   InventoryVM,
   InventoryItemVM,
-  EncumbranceVM,
-  EncumbranceTier,
   CoinVM,
 } from "@domain/vm-types";
 import { FLAGS, readFlag } from "@domain/flags";
@@ -329,189 +326,9 @@ export function sortEquipped(items: InventoryItemVM[]): InventoryItemVM[] {
   );
 }
 
-// ---------------------------------------------------------------------------
-// Encumbrance
-// ---------------------------------------------------------------------------
-
-// Status by tier — indexed by EncumbranceTier. OSE has no localized labels for
-// weight tiers, so these are ours; tier itself comes from the system's breakpoints.
-const TIER_STATUS = [
-  "Unencumbered",
-  "Lightly encumbered",
-  "Heavily encumbered",
-  "Severely encumbered",
-  "Overloaded",
-] as const;
-
-/** OSE's default significant-treasure threshold, in cn (basic encumbrance). */
-export const DEFAULT_SIGNIFICANT_TREASURE = 800;
-
-/**
- * World significant-treasure setting (basic encumbrance), in cn. OSE registers it as a
- * NUMBER — coercing it to a boolean fed `true` into the encumbrance ctor, where
- * `options.significantTreasure || 800` kept the `true` and made every threshold check
- * `value >= true`, i.e. tripping at 1cn of treasure. Safe in non-Foundry tests.
- */
-export function significantTreasure(): number {
-  try {
-    const settings = game.settings as { get(ns: string, key: string): unknown };
-    const v = Number(settings.get(game.system.id, "significantTreasure"));
-    return Number.isFinite(v) && v > 0 ? v : DEFAULT_SIGNIFICANT_TREASURE;
-  } catch {
-    return DEFAULT_SIGNIFICANT_TREASURE;
-  }
-}
-
-/**
- * Constructor shape of OSE's active encumbrance class (variant per world setting).
- * The `ose` types package exports the CharacterEncumbrance *instance* interface but
- * not the variant classes (nor a matching ctor sig), so the ctor is spelled here.
- */
-type EncumbranceCtor = new (
-  max: number,
-  items: OseItem[],
-  options: { significantTreasure: number; scores: unknown },
-  strMod: number,
-) => CharacterEncumbrance;
-
-export function selectEncumbrance(
-  actor: OSEActor,
-  items?: OseItem[],
-): EncumbranceVM {
-  const live = actor.system.encumbrance;
-  // A partial/mock actor may lack encumbrance entirely — fall back to a disabled
-  // readout (moveBands still read off actor.system.movement below) rather than crash.
-  if (!live) {
-    const movement = actor.system.movement;
-    return {
-      enabled: false,
-      value: 0,
-      max: 0,
-      pct: 0,
-      tier: 0,
-      status: TIER_STATUS[0],
-      label: "",
-      moveBands: {
-        encounter: movement?.encounter ?? 0,
-        explore: movement?.base ?? 0,
-        travel: movement?.overland ?? 0,
-      },
-      bands: [],
-    };
-  }
-  // Optimistically recompute from the overlaid items via OSE's OWN encumbrance class,
-  // mirroring the system's construction, so equip/nest/consume/qty changes reflect
-  // instantly instead of waiting for the actor to re-prep. `max` (STR-derived cap) is
-  // item-independent so it's taken off the live instance. No items ⇒ live instance (tests).
-  const e =
-    items === undefined
-      ? live
-      : new (live.constructor as unknown as EncumbranceCtor)(
-          live.max,
-          items,
-          {
-            significantTreasure: significantTreasure(),
-            scores: actor.system.scores,
-          },
-          actor.system.scores?.str?.mod ?? 0,
-        );
-  const movement = actor.system.movement;
-  const sigTreasure = significantTreasure();
-  // Basic encumbrance gauges carried TREASURE (`e.value` in basic is exactly the treasure cn,
-  // coins included) against the 1600-cn immobile cap (`e.max`): the bar fills value/max, reads
-  // green below the significant-treasure line (`steps[0]`, ~50% at defaults), then ramps
-  // yellow→red as the load nears the cap (solid red once immobile). Weight/slot variants keep
-  // their own value/max fill + system breakpoint bands.
-  const immobile = e.variant === "basic" && e.value >= e.max;
-  const pct = e.max > 0 ? Math.min(1, e.value / e.max) : 0;
-  // Colour breakpoints (% of cap). The SAME list drives the bar's colour transitions AND the
-  // tier that tints the movement scores, so the two can never disagree. Basic OSE exposes a
-  // single step (the treasure line); we add two visual stops so the fill escalates toward red
-  // instead of holding flat yellow up to the cap.
-  const steps = immobile ? [] : (e.steps ?? []);
-  const bands =
-    e.variant === "basic" && steps.length === 1
-      ? [steps[0], steps[0] + (100 - steps[0]) / 2, 90]
-      : steps;
-  // Tier = the band the fill sits in — NOT armor-driven. Armor slows movement in basic OSE
-  // but must not tint the bar/scores; it surfaces as its own popover row (see armorTier).
-  const overloaded = immobile || e.encumbered;
-  const tier = fillTier(pct, bands, overloaded);
-  const unit = e.variant === "itembased" ? "items" : "cn";
-  return {
-    enabled: e.enabled,
-    value: e.value,
-    max: e.max,
-    pct,
-    tier,
-    status: TIER_STATUS[tier],
-    label: `${e.value} / ${e.max} ${unit}`,
-    armorTier: basicArmorTier(e, sigTreasure),
-    moveBands: {
-      encounter: movement?.encounter ?? 0,
-      explore: movement?.base ?? 0,
-      travel: movement?.overland ?? 0,
-    },
-    bands,
-    // Solid-fill colour when bands is empty (basic-immobile → red enc-4).
-    barTier: tier,
-  };
-}
-
-/**
- * Colour tier for a fill: how many breakpoints `pct` (0–1) has crossed, clamped 0–4.
- * `overloaded` (over the immobile cap, or the system's `encumbered` flag) forces the top
- * tier. Bar transitions and movement-score colour both read from this, so they agree.
- */
-function fillTier(pct: number, bands: number[], overloaded: boolean): EncumbranceTier {
-  if (overloaded) return 4;
-  const p = pct * 100;
-  const crossed = bands.reduce((n, b) => (p >= b ? n + 1 : n), 0);
-  return Math.min(crossed, 4) as EncumbranceTier;
-}
-
-/**
- * Basic-mode armour tier, recovered from the system's breakpoint flags + treasure state.
- * Armour slows movement in basic OSE but must NOT tint the bar/scores — this drives a
- * dedicated popover row instead. Undefined for non-basic variants.
- */
-function basicArmorTier(
-  e: CharacterEncumbrance,
-  sigTreasure: number,
-): "unarmored" | "light" | "heavy" | undefined {
-  if (e.variant !== "basic") return undefined;
-  const heavy = e.atThirdBreakpoint || (e.atSecondBreakpoint && !e.atFirstBreakpoint);
-  if (heavy) return "heavy";
-  const overTreasure = e.value >= sigTreasure;
-  const light = e.atSecondBreakpoint || (e.atFirstBreakpoint && !overTreasure);
-  return light ? "light" : "unarmored";
-}
-
-/** Half-width (in %) of the colour fade centred on each tier threshold. */
-const ENC_FADE_HALF = 4;
-
-/**
- * CSS colour-stop list for the encumbrance bar. Each tier holds its pure `--enc-N`
- * colour across its band, then fades into the next over a narrow window CENTRED on
- * the real threshold — so at a boundary % the colour is exactly 50/50 between the two
- * tiers (the tier change still reads at the right spot), but the segments blend
- * smoothly instead of butting hard edges. No thresholds ⇒ a solid bar in `barTier`'s
- * colour (falling back to `tier`). Colours are the shared `--enc-N` tokens — one tier, one colour.
- */
-export function encBarStops({
-  bands,
-  tier,
-  barTier,
-}: Pick<EncumbranceVM, "bands" | "tier" | "barTier">): string {
-  if (bands.length === 0) return `var(--enc-${barTier ?? tier}) 0 100%`;
-  const stops = ["var(--enc-0) 0%"];
-  bands.forEach((t, i) => {
-    const from = Math.max(0, t - ENC_FADE_HALF);
-    const to = Math.min(100, t + ENC_FADE_HALF);
-    // hold this tier's colour up to the fade's start, then ramp to the next by its end
-    stops.push(`var(--enc-${Math.min(i, 4)}) ${from}%`);
-    stops.push(`var(--enc-${Math.min(i + 1, 4)}) ${to}%`);
-  });
-  stops.push(`var(--enc-${Math.min(bands.length, 4)}) 100%`);
-  return stops.join(", ");
-}
+export {
+  DEFAULT_SIGNIFICANT_TREASURE,
+  encBarStops,
+  selectEncumbrance,
+  significantTreasure,
+} from "./encumbrance";
